Add support for 7.5" waveshare screen, version 2 (800x480 display and support for fast update)
