+++ .gitignore
@@ -1,0 +1,3 @@
+*.o
+*.swp
+ultimatewriter
